Rename YourEmail -> YourInfo
It's about to include name, too, so the leader can be joined to the
seder right away
