+++ frontend-create-haggadah/src/YourInfoSection.tsx
